Tried MACVTAP interface for VMs
MACVTAP entrypoint can be found in veos directory as entrypoint.sh.macvtap. With MACVTAP everything works, except for the fact that /dev/tap interfaces are created outside of container and are not accessible inside. It also seems like they don't get deleted after the macvtap interface goes (as I can see interface index on my docker host increasing all the time during tests). So all in all it looks dodgy and a potential headache. For now i'm leaving the standard bridge networking in place, will need to monitor the bug referenced in README