Support latest image tag for latest images
Code mentions
```
This assumes that docker image tag contains the correct EOS version, e.g. ceos:4.20.5
```

but in practice all the examples provided here contain "latest" as the EOS version. docker-topo will fail to start containers with recent images tagged as latest as the command line is incorrect, so we fix this specific case here.